requirements.txt
flask==2.3.3
python-dotenv==1.0.0
pytest==7.4.0 
supabase==2.13.0
numpy==2.2.3
flask-cors==4.0.0
gunicorn==21.2.0